requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
py==1.11.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
retry==0.9.2
retrying==1.3.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0